web/requirements.txt
amqp==2.5.2
billiard==3.6.2.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.0
importlib-metadata==1.0.0
kombu==4.6.6
more-itertools==8.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
